pyproject.toml
[project]
name = "mcp-karpenter-documentation"
version = "0.1.0"
description = "MCP server for searching and reading AWS Karpenter documentation"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "martoc" },
]
readme = "README.md"
keywords = ["mcp", "karpenter", "documentation", "aws", "kubernetes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
mcp-karpenter-documentation = "mcp_karpenter_documentation.server:run_server"
karpenter-docs-index = "mcp_karpenter_documentation.cli:main"

[project.urls]
homepage = "https://github.com/martoc/mcp-karpenter-documentation"
repository = "https://github.com/martoc/mcp-karpenter-documentation"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_karpenter_documentation"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[tool.uv]
required-version = ">=0.5.0"

[tool.pytest.ini_options]
addopts = "-vvv --cov=src/mcp_karpenter_documentation --cov-report=term-missing --cov-report=html:target/coverage"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
extend-exclude = [
    "__pycache__",
    "build",
    "dist",
    "target",
    ".venv",
]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "D",      # pydocstyle
    "N",      # pep8-naming
    "S",      # flake8-bandit (security)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "RUF",    # Ruff-specific rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "D"]

[tool.mypy]
files = ["src", "tests"]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
